webserver.resource-manager settings
